+++ packages/manager/modules/pci/src/projects/new/component.js
@@ -11,5 +11,6 @@ export default {
     summary: '<',
     getSummary: '<',
     trackProjectCreationError: '<',
+    model: '<',
   },
 };