+++ packages/manager/modules/pci/src/projects/new/components/pci-new-project-promotion-banner/component.js
@@ -1,0 +1,12 @@
+import controller from './controller';
+import template from './template.html';
+
+const component = {
+  template,
+  controller,
+  bindings: {
+    voucher: '<',
+  },
+};
+
+export default component;